% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Test hello world function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the object to say hi to}
}
\value{
The printed output \code{\link{print}}
}
\description{
Test hello world function
}
\examples{
hello("Alex")
}
